city done

+++ src/main/java/example/addressrepo/service/AddressService.java
@@ -50,6 +50,9 @@ public class AddressService {
     }
 
     public void deleteAddress(UUID addressId) {
+        if (!addressRepository.existsById(addressId)) {
+            throw new DoesNotExistsException("Address with Id: " + addressId + " does not exist");
+        }
         addressRepository.deleteById(addressId);
     }
